Compare Column Names Between Two data.tables

This function compares the column names of two data.table objects and identifies which columns are missing in each.

Usage

compare_colnames(dt1, dt2, return_list = FALSE)

Arguments

dt1

A data.table object.

dt2

A data.table object.

return_list

Logical. If TRUE, the function returns a list with missing columns; if FALSE, it returns a formatted kable table. Default is FALSE.

Value

If return_list = TRUE, a named list with:

missing_in_dt1

Columns present in dt2 but missing in dt1

missing_in_dt2

Columns present in dt1 but missing in dt2

If return_list = FALSE, a kable table showing missing columns.

Examples

library(data.table)

dt1 = data.table(a = 1:5, b = 6:10, c = 11:15)
dt2 = data.table(b = 1:5, c = 6:10, d = 11:15)

# Return a kable table
compare_colnames(dt1, dt2)
#> 
#> 
#> |Column |Missing_In |
#> |:------|:----------|
#> |d      |dt1        |
#> |a      |dt2        |

# Return a list with missing columns
compare_colnames(dt1, dt2, return_list = TRUE)
#> [[1]]
#> [[1]]$missing_in_dt1
#> [1] "d"
#> 
#> 
#> [[2]]
#> [[2]]$missing_in_dt2
#> [1] "a"
#> 
#>