Retrieve and Process ENTSO-E Actual Generation Data

This function retrieves and processes the ENTSO-E actual generation data for a specified year and month. It supports an option to either retain or delete the raw CSV file after processing based on the raw argument. It also allows specifying the output folder for the downloaded CSV file.

entsoe_actual_generation(
  year_data,
  month_data,
  raw = FALSE,
  user = Sys.getenv("ENTSOE_USER"),
  password = Sys.getenv("ENTSOE_PASSWORD"),
  output_folder = tempdir()
)

Arguments

year_data

Numeric. The year for which to download the ENTSO-E data (e.g., 2024).

month_data

Numeric. The month for which to download the ENTSO-E data (e.g., 10 for October).

raw

Logical. If TRUE, keeps the downloaded CSV file after processing. If FALSE, the CSV file is deleted after being read. Default is FALSE.

user

Character. Your ENTSO-E API key. Passed to entsoe_download_file for authenticating the download request. Default is Sys.getenv('ENTSOE_USER').

password

Character. Your ENTSO-E API key. Passed to entsoe_download_file for authenticating the download request. Default is Sys.getenv('ENTSOE_PASSWORD').

output_folder

Character. The folder where the raw CSV file will be saved. Default is tempdir().

Value

A data.table in long format with the following columns:

DATE

The date of the data point.

TIME

The time of the data point in HH:MM format.

HOUR

The hour extracted from the DATETIME column as an integer.

RESOLUTION

The resolution code (e.g., quarter-hour, hourly).

CODE_MAP

The map code (e.g., country, bidding zone).

CODE_EIC

The EIC code for the area (e.g., grid or market region).

PRODUCTION_TYPE

The type of power generation (e.g., "Fossil Oil", "Wind Onshore").

ASSET_CATEGORY

The category of the asset (e.g., "Thermal", "Wind", etc.).

ACTUAL_GENERATION

The actual generation output for the given time period.

ACTUAL_CONSUMPTION

The actual consumption for the given time period.

UPDATETIME

The last update time for the data entry.

VARIABLE

The variable being measured, either "ACTUAL_GENERATION" or "ACTUAL_CONSUMPTION".

VALUE

The value for the corresponding variable.

Examples

if (FALSE) { # \dontrun{
# Download and process the data for October 2024, keeping the CSV file
result_raw <- entsoe_actual_generation(2024, 10, raw = TRUE)

# Download and process the data for October 2024, deleting the CSV file after processing
result <- entsoe_actual_generation(2024, 10)
} # }