Retrieve a Specific File from ENTSO-E SFTP Server

This function requests and retrieves a specific CSV file from the ENTSO-E SFTP server based on a folder (basis_name), year, and month. It authenticates using credentials stored in environment variables (ENTSOE_USER and ENTSOE_PASSWORD).

Usage

entsoe_create_url_file(basis_name, year, month)

Arguments

basis_name

A character string specifying the folder name within the ENTSO-E SFTP server's transparency platform.

year

A numeric or character string specifying the year of the file to be retrieved.

month

A numeric or character string specifying the month of the file to be retrieved.

Value

A response object containing the contents of the specified file.

Details

The function constructs the file path from basis_name, year, and month and uses curl to retrieve the specified CSV file from the server.