Retrieve Day-Ahead Market Prices for a Specific Country

This function retrieves Day-Ahead Market (DAM) prices for a specific country from the ENTSOE Transparency Platform API. The function fetches market price data for a given time range and country, processes the data, and formats it into a data.table.

This function retrieves Day-Ahead Market (DAM) prices for a specific country from the ENTSOE Transparency Platform API. The function fetches market price data for a given time range and country, processes the data, and formats it into a data.table.

Usage

entsoe_dam_prices(
  country,
  from_data,
  to_data,
  api_key = Sys.getenv("ENTSOE_KEY"),
  verbose = FALSE,
  plot = FALSE
)

Arguments

country

A character string specifying the country for which to retrieve DAM prices. The country should match an entry in the entsoe_countries table.

from_data

A Date object specifying the start date for the data range. Default is 365 days before the current date.

to_data

A Date object specifying the end date for the data range. Default is the current date.

api_key

A character string containing the ENTSOE API key. It can be passed directly or retrieved from the environment variable ENTSOE_KEY.

Value

A data.table containing the formatted DAM price data for the specified country and time range. The data table includes:

DATE: The date in YYYY-MM-DD format.

HOUR: The hour of the day (position in the data).

VALUE: The price value in EUR.

UNIT: The unit of the price (always 'EUR').

A data.table containing the formatted DAM price data for the specified country and time range. The data table includes:

DATE: The date in YYYY-MM-DD format.

HOUR: The hour of the day (position in the data).

VALUE: The price value in EUR.

UNIT: The unit of the price (always 'EUR').

Examples

# Example usage to retrieve DAM prices for Italy (North)
country <- "Italy (North)"
from_data <- Sys.Date() - 365
to_data <- Sys.Date()
dam_prices <- entsoe_dam_prices(country, from_data, to_data, api_key = Sys.getenv('ENTSOE_KEY'))
#> [OK] DATA for Italy (North) from 2024-04-01 to 2024-04-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-05-01 to 2024-05-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-06-01 to 2024-06-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-07-01 to 2024-07-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-08-01 to 2024-08-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-09-01 to 2024-09-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-10-01 to 2024-10-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-11-01 to 2024-11-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-12-01 to 2024-12-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-01-01 to 2025-01-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-02-01 to 2025-02-28 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-03-01 to 2025-03-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-04-01 to 2025-04-28 retrieved correctly 
print(dam_prices)
#>       COUNTRY   CODE_ENTSOE       DATE  HOUR RESOLUTION  VALUE   UNIT
#>        <char>        <char>     <char> <int>     <char>  <num> <char>
#>    1:   Italy Italy (North) 2025-03-30     2      PT15M 126.00    EUR
#>    2:   Italy Italy (North) 2024-04-01     1      PT60M  90.84    EUR
#>    3:   Italy Italy (North) 2024-04-01     2      PT60M  90.00    EUR
#>    4:   Italy Italy (North) 2024-04-01     3      PT60M  87.67    EUR
#>    5:   Italy Italy (North) 2024-04-01     4      PT60M  85.67    EUR
#>   ---                                                                
#> 9596:   Italy Italy (North) 2025-04-28    20      PT15M  91.95    EUR
#> 9597:   Italy Italy (North) 2025-04-28    21      PT15M  97.10    EUR
#> 9598:   Italy Italy (North) 2025-04-28    22      PT15M  97.10    EUR
#> 9599:   Italy Italy (North) 2025-04-28    23      PT15M  97.10    EUR
#> 9600:   Italy Italy (North) 2025-04-28    24      PT15M  97.10    EUR
#>                  DATETIME
#>                    <POSc>
#>    1:                <NA>
#>    2: 2024-04-01 01:00:00
#>    3: 2024-04-01 02:00:00
#>    4: 2024-04-01 03:00:00
#>    5: 2024-04-01 04:00:00
#>   ---                    
#> 9596: 2025-04-28 20:00:00
#> 9597: 2025-04-28 21:00:00
#> 9598: 2025-04-28 22:00:00
#> 9599: 2025-04-28 23:00:00
#> 9600: 2025-04-29 00:00:00

# Example usage to retrieve DAM prices for Italy (North)
country <- "Italy (North)"
from_data <- Sys.Date() - 365
to_data <- Sys.Date()
dam_prices <- entsoe_dam_prices(country, from_data, to_data, api_key = Sys.getenv('ENTSOE_KEY'))
#> [OK] DATA for Italy (North) from 2024-04-01 to 2024-04-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-05-01 to 2024-05-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-06-01 to 2024-06-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-07-01 to 2024-07-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-08-01 to 2024-08-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-09-01 to 2024-09-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-10-01 to 2024-10-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-11-01 to 2024-11-30 retrieved correctly 
#> [OK] DATA for Italy (North) from 2024-12-01 to 2024-12-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-01-01 to 2025-01-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-02-01 to 2025-02-28 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-03-01 to 2025-03-31 retrieved correctly 
#> [OK] DATA for Italy (North) from 2025-04-01 to 2025-04-28 retrieved correctly 
print(dam_prices)
#>       COUNTRY   CODE_ENTSOE       DATE  HOUR RESOLUTION  VALUE   UNIT
#>        <char>        <char>     <char> <int>     <char>  <num> <char>
#>    1:   Italy Italy (North) 2025-03-30     2      PT15M 126.00    EUR
#>    2:   Italy Italy (North) 2024-04-01     1      PT60M  90.84    EUR
#>    3:   Italy Italy (North) 2024-04-01     2      PT60M  90.00    EUR
#>    4:   Italy Italy (North) 2024-04-01     3      PT60M  87.67    EUR
#>    5:   Italy Italy (North) 2024-04-01     4      PT60M  85.67    EUR
#>   ---                                                                
#> 9596:   Italy Italy (North) 2025-04-28    20      PT15M  91.95    EUR
#> 9597:   Italy Italy (North) 2025-04-28    21      PT15M  97.10    EUR
#> 9598:   Italy Italy (North) 2025-04-28    22      PT15M  97.10    EUR
#> 9599:   Italy Italy (North) 2025-04-28    23      PT15M  97.10    EUR
#> 9600:   Italy Italy (North) 2025-04-28    24      PT15M  97.10    EUR
#>                  DATETIME
#>                    <POSc>
#>    1:                <NA>
#>    2: 2024-04-01 01:00:00
#>    3: 2024-04-01 02:00:00
#>    4: 2024-04-01 03:00:00
#>    5: 2024-04-01 04:00:00
#>   ---                    
#> 9596: 2025-04-28 20:00:00
#> 9597: 2025-04-28 21:00:00
#> 9598: 2025-04-28 22:00:00
#> 9599: 2025-04-28 23:00:00
#> 9600: 2025-04-29 00:00:00