Retrieve Actual Generation by Production Type for a Specific Country

This function retrieves actual generation by production type data for a specific country from the ENTSOE Transparency Platform API. The function fetches generation data for a given time range and country, processes the data, and formats it into a data.table.

Usage

entsoe_generation_by_type(
  country,
  from_data,
  to_data,
  api_key = Sys.getenv("ENTSOE_KEY")
)

Arguments

country

A character string specifying the country for which to retrieve generation data. The country should match an entry in the entsoe_countries table.

from_data

A Date object specifying the start date for the data range. Default is 365 days before the current date.

to_data

A Date object specifying the end date for the data range. Default is the current date.

api_key

A character string containing the ENTSOE API key. It can be passed directly or retrieved from the environment variable ENTSOE_KEY.

Value

A data.table containing the formatted generation data for the specified country and time range. The data table includes:

DATE: The date in YYYY-MM-DD format.

HOUR: The hour of the day (position in the data).

PRODUCTION_TYPE: The type of production (e.g., solar, wind, nuclear).

VALUE: The generation value in MW.

UNIT: The unit of the generation value (always 'MW').

Examples

# Example usage to retrieve generation data for Italy (North)
country <- "Italy (North)"
from_data <- Sys.Date() - 365
to_data <- Sys.Date()
generation_data <- entsoe_generation_by_type(country, from_data, to_data, api_key = Sys.getenv('ENTSOE_KEY'))
#> https://web-api.tp.entsoe.eu/api?documentType=A75&processType=A16&in_Domain=10Y1001A1001A73I&periodStart=202404280000&periodEnd=202504290000&securityToken=<...>
#> response has arrived
#> [OK] DATA for Italy (North) from 2024-04-28 to 2025-04-29 retrieved correctly 
print(generation_data)
#>                        date position value    uom             zone
#>                      <POSc>    <num> <num> <char>           <char>
#>      1: 2024-04-28 00:00:00        1   153    MAW 10Y1001A1001A73I
#>      2: 2024-04-28 01:00:00        2   153    MAW 10Y1001A1001A73I
#>      3: 2024-04-28 02:00:00        3   153    MAW 10Y1001A1001A73I
#>      4: 2024-04-28 03:00:00        4   151    MAW 10Y1001A1001A73I
#>      5: 2024-04-28 04:00:00        5   152    MAW 10Y1001A1001A73I
#>     ---                                                           
#> 210036: 2025-04-28 05:45:00       32    17    MAW 10Y1001A1001A73I
#> 210037: 2025-04-28 06:00:00       33    19    MAW 10Y1001A1001A73I
#> 210038: 2025-04-28 06:15:00       34    16    MAW 10Y1001A1001A73I
#> 210039: 2025-04-28 06:30:00       35    15    MAW 10Y1001A1001A73I
#> 210040: 2025-04-28 06:45:00       36    10    MAW 10Y1001A1001A73I
#>                           data COUNTRY   CODE_ENTSOE
#>                         <char>  <char>        <char>
#>      1: actual_generation_type   Italy Italy (North)
#>      2: actual_generation_type   Italy Italy (North)
#>      3: actual_generation_type   Italy Italy (North)
#>      4: actual_generation_type   Italy Italy (North)
#>      5: actual_generation_type   Italy Italy (North)
#>     ---                                             
#> 210036: actual_generation_type   Italy Italy (North)
#> 210037: actual_generation_type   Italy Italy (North)
#> 210038: actual_generation_type   Italy Italy (North)
#> 210039: actual_generation_type   Italy Italy (North)
#> 210040: actual_generation_type   Italy Italy (North)