List Files within a Specified Folder on ENTSO-E SFTP Server

This function retrieves and lists files from a specified folder (basis_name) on the ENTSO-E SFTP server's transparency platform. It uses the entsoe_create_url_files function to fetch the file listing and processes the raw content with parse_ftp_files.

entsoe_list_files(basis_name)

Arguments

basis_name

A character string specifying the name of the folder from which to retrieve the file listing.

Value

A data frame containing information about files within the specified folder on the ENTSO-E SFTP server. The data frame includes file names and any other available metadata.

Details

The function sends a request to the ENTSO-E SFTP server, retrieves the list of files as raw text, and converts it into a structured format using parse_ftp_files. Ensure that basis_name corresponds to an existing folder on the server.

Examples

if (FALSE) { # \dontrun{
# List files in a folder named "example_folder":
head(entsoe_list_files(basis_name = "OutagesPU"))
} # }