Validate Filename Structure

This function checks whether a given filename matches a specific structure based on the number of leading digits, a middle expression, and the file extension.

gme_validate_filename(filename, num_digits = 8, file_extension = "xml")

Arguments

filename

A string representing the filename to validate.

num_digits

An integer specifying the number of leading digits in the filename. Default is 8.

file_extension

A string specifying the expected file extension (without the dot). Default is "xml".

middle_expression

A string specifying the exact middle expression in the filename. Default is "MGPPrezzi".

Value

A logical value (TRUE if the filename is valid, FALSE otherwise). If the filename is invalid, it prints "Wrong filename" to the console.

Examples

# Validate a correctly formatted filename
gme_validate_filename("20241126MGPPrezzi.xml") # Returns TRUE
#> [1] TRUE

# Validate an incorrectly formatted filename
gme_validate_filename("wrongfile.xml") # Prints "Wrong filename" and returns FALSE
#> Wrong filename 
#> 
#> [1] FALSE

# Validate with custom parameters
gme_validate_filename("2024MGPPrezzi.txt", num_digits = 4, file_extension = "txt") # Returns TRUE
#> [1] TRUE