Process Excel file and extract relevant data

This function processes an Excel file containing multiple sheets, extracts relevant data based on the type of commodity (gas or electricity), filters the data within the specified date range, and combines it into a single data.table. The data is reshaped, cleaned, and additional columns are added for further analysis.

process_alba_xlsx(filename, from_date, to_date, type)

Arguments

filename

A character string specifying the path to the Excel file to be processed.

from_date

A Date object or a character string in "YYYY-MM-DD" format specifying the start date for filtering the data.

to_date

A Date object or a character string in "YYYY-MM-DD" format specifying the end date for filtering the data.

type

A character string specifying the type of commodity. It can be either "gas" or "electricity".

Value

A data.table containing the processed data, with columns:

DATE: The date of the observation.

TYPE: The commodity type (e.g., "PSV", "COAL").

VARIABLE: The name of the variable (e.g., "PSV MW").

VALUE: The value of the variable for the given date.

Details

This function processes the following sheets for "gas" type: "PSV MW", "GR04 MW", "TTF MW", "GR07 MW", "Baum MW", "NCG MW", "PEG MW". For "electricity" type, the following sheets are processed: "Elettricita Italia", "Elettricita Francia", "Elettricita Germania", "COAL", and "CO2". Data is filtered by the from_date and to_date parameters, reshaped from wide to long format, and combined into a single data.table.

Examples

# Example usage:
filename <- "path/to/your/file.xlsx"
from_date <- Sys.Date() - 7
to_date <- Sys.Date()
type <- "electricity"
processed_data <- process_xlsx(filename, from_date, to_date, type)
#> Error in process_xlsx(filename, from_date, to_date, type): could not find function "process_xlsx"