Make an API Request to ENTSO-E Web API

This function sends a request to the ENTSO-E (European Network of Transmission System Operators for Electricity) Web API, retrieves data based on the given query, and processes the response. It handles both XML and zip file responses.

Usage

api_req(
  api_scheme = "https://",
  api_domain = "web-api.tp.entsoe.eu/",
  api_name = "api?",
  query_string = NULL,
  api_key = NULL
)

Arguments

api_scheme

A string representing the API scheme, default is "https://". This is the protocol used for the API request.

api_domain

A string representing the domain name of the API, default is "web-api.tp.entsoe.eu/".

api_name

A string representing the API name or endpoint, default is "api?".

query_string

A string representing the query parameters to be appended to the API URL. This parameter is required.

api_key

A string representing the security token used for authentication. This parameter is required.

Value

A list containing the response data:

en_cont_list

A list of XML content if the response is a zip file. The list contains the decompressed XML files.

en_cont

XML content if the response is an XML document.

Details

The function first validates the presence of the required arguments: query_string and api_key. If the API request is successful, it handles both XML responses and zip file responses:

If the response is a zip file (application/zip), it downloads and extracts the XML content.

If the response is XML (text/xml or application/xml), it returns the XML content directly.

If the request fails with a 200 OK response but contains an error message or exceeds the data limit, the function checks for the need to offset the query and recursively retries with a new query offset.

Examples

if (FALSE) { # \dontrun{
# Example usage of the api_req function
query_string <- "documentType=A63&businessType=A85"
api_key <- "your_api_key"
result <- api_req(query_string = query_string, api_key = api_key)
print(result)
} # }