Download and Process MGP Data File

This function downloads a file from an FTP server using provided credentials, processes the XML data from the downloaded file, and returns the processed data. After processing, the downloaded XML file is deleted.

Usage

mgp_download_file(
  filename,
  data_type = "MGP_Prezzi",
  output_dir,
  username,
  password,
  raw = FALSE
)

Arguments

filename

A character string representing the name of the file to be downloaded.

output_dir

A character string representing the directory where the downloaded file will be saved.

username

A character string representing the FTP username for authentication.

password

A character string representing the FTP password for authentication.

Value

A data frame (or NULL in case of an error). The data frame contains the processed data obtained from the downloaded XML file, or NULL if an error occurred during download or processing.

Details

This function uses the curl package to download the file from an FTP server with the provided credentials. After downloading, it passes the downloaded file to the gme_dam_xml_to_data function for processing. Finally, the downloaded XML file is deleted from the local system to clean up.

Examples

# Example usage:
result <- mgp_download_file("example.xml", "your_username", "your_password", "/path/to/output_dir")
#> Error in mgp_download_file("example.xml", "your_username", "your_password",     "/path/to/output_dir"): Invalid data_type: your_username . Must be one of: MGP_Prezzi, MGP_Quantita, MGP_Fabbisogno, MGP_Liquidita, MGP_Transiti, MGP_LimitiTransito
print(result)
#> Error: object 'result' not found